+++ src/components/DatePicker/BsDatePicker.ts
@@ -16,17 +16,17 @@ export default defineComponent<TBsDatePicker, TRecord, TRecord, ComputedOptions,
     ],
     setup(props, {emit}) {
         const thisProps = props as Readonly<TDatePickerOptionProps>;
-        const locale = ref<string>(thisProps.locale || (isServer ? 'en-US' : window.navigator.language));
+        const locale = ref<string>(thisProps.locale || (isServer ? DatePickerConst.defaultLocale : window.navigator.language));
         const localValue = ref(useParseDate(thisProps.modelValue).setLocale(locale.value));
         const calendarDate = ref<Date>(localValue.value.toJSDate());
         const currentView = ref<TDateTimePickerMode>(
-            <TDateTimePickerMode>(thisProps.viewMode || thisProps.mode || 'date')
+            <TDateTimePickerMode>(thisProps.viewMode || thisProps.mode || DatePickerConst.DATE)
         );
         const pickerMode = computed(() =>
             <TDateTimePickerMode>(thisProps.viewMode || thisProps.mode || DatePickerConst.DATE)
         );
         const showTime = computed(() =>
-            ['datetime', 'time'].includes(pickerMode.value)
+            [DatePickerConst.DATETIME, DatePickerConst.TIME].includes(pickerMode.value)
         );
         const ensureViewMode = () => {
             if (currentView.value === DatePickerConst.DATETIME) {